% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wt.R
\name{prev_wt}
\alias{prev_wt}
\title{Weighted prevalence - adjusted for nonresponse
Calcuated weighted prevalence for a given variable}
\usage{
prev_wt(data, wt_var, subpop = NULL)
}
\arguments{
\item{data}{dataset. must contain strata, permid, wt_n, wt_Nc}

\item{wt_var}{variable to estimate, adjusting for nonresponse. input as character}

\item{subpop}{subpopulation}
}
\value{
data frame
}
\description{
Weighted prevalence - adjusted for nonresponse
Calcuated weighted prevalence for a given variable
}
\examples{
\dontrun{
df = read_dta("data/ddis2_hw.dta")
df = data.frame(df)
prev_wt(data=df, wt_var="anypdsm_t0")
}
}
\author{
David Aaby <david.aaby@northwestern.edu>
}
