% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_corr.R
\name{wt_corr}
\alias{wt_corr}
\title{Survey Weighting - adjustment for nonresponse
Adjust the baseline NJP weights for nonresponse in a particular variable}
\usage{
wt_corr(data, wt_var)
}
\arguments{
\item{data}{dataset. must contain strata, permid, wt_n, wt_Nc}

\item{wt_var}{variable to weight, adjusting for nonresponse. input as character}
}
\value{
An object of class \code{survey.design}
}
\description{
Survey Weighting - adjustment for nonresponse
Adjust the baseline NJP weights for nonresponse in a particular variable
}
\note{
This code was tranlasted from Stata wt_corr.ado, written by GMAC  For example, you can set alpha=0.10 to obtain one-sided test at 0.05 significance level.
}
\examples{
\dontrun{
df = read_dta("data/ddis2_hw.dta")
df = data.frame(df)
wt_corr(data=df, wt_var="anypdsm_t0")
}
}
\author{
David Aaby <david.aaby@northwestern.edu>
}
